no longer uses django models
Attempting to make website so that it does not need to download the whole table to search through it. It still technically does that, but doesn't save to a file like before.

Models were getting a bit of a hassle. I suppose I could have used .object.get_or_create() but instead it gets lists that do not get saved when searching